requirements.txt
python-dotenv
py-trello
pandas
openpyxl